refactor: group files by folders
