+++ src/blocks/HomePageBlocks/CardStackSection.tsx
