+++ src/components/ApplicationLogo/ApplicationLogo.tsx
