+++ src/components/AuthCard/AuthCard.tsx
