+++ src/pages/dashboard.tsx
@@ -1,11 +1,10 @@
-import AppLayout from '@/components/Layouts/AppLayout';
+import AppLayout from '@/layouts/AppLayout';
 import Head from 'next/head';
 
 const Dashboard = () => {
   return (
     <AppLayout
-      header={<h2 className="font-semibold text-xl text-gray-800 leading-tight">Dashboard</h2>}
-    >
+      header={<h2 className="font-semibold text-xl text-gray-800 leading-tight">Dashboard</h2>}>
       <Head>
         <title>Laravel - Dashboard</title>
       </Head>
