+++ src/pages/forgot-password.tsx
@@ -1,11 +1,11 @@
-import ApplicationLogo from '@/components/ApplicationLogo';
-import AuthCard from '@/components/AuthCard';
-import AuthSessionStatus from '@/components/AuthSessionStatus';
-import Button from '@/components/Button';
-import GuestLayout from '@/components/Layouts/GuestLayout';
-import Input from '@/components/Input';
-import InputError from '@/components/InputError';
-import Label from '@/components/Label';
+import ApplicationLogo from '@/components/ApplicationLogo/ApplicationLogo';
+import AuthCard from '@/components/AuthCard/AuthCard';
+import AuthSessionStatus from '@/components/AuthSessionStatus/AuthSessionStatus';
+import Button from '@/components/Button/Button';
+import GuestLayout from '@/layouts/GuestLayout';
+import Input from '@/components/Input/Input';
+import InputError from '@/components/InputError/InputError';
+import Label from '@/components/Label/Label';
 import Link from 'next/link';
 import { useAuth } from '@/hooks/auth';
 import { useState } from 'react';
@@ -34,8 +34,7 @@ const ForgotPassword = () => {
               <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
             </a>
           </Link>
-        }
-      >
+        }>
         <div className="mb-4 text-sm text-gray-600">
           Forgot your password? No problem. Just let us know your email address and we will email
           you a password reset link that will allow you to choose a new one.
