+++ src/pages/login.tsx
@@ -1,11 +1,11 @@
-import ApplicationLogo from '@/components/ApplicationLogo';
-import AuthCard from '@/components/AuthCard';
-import AuthSessionStatus from '@/components/AuthSessionStatus';
-import Button from '@/components/Button';
-import GuestLayout from '@/components/Layouts/GuestLayout';
-import Input from '@/components/Input';
-import InputError from '@/components/InputError';
-import Label from '@/components/Label';
+import ApplicationLogo from '@/components/ApplicationLogo/ApplicationLogo';
+import AuthCard from '@/components/AuthCard/AuthCard';
+import AuthSessionStatus from '@/components/AuthSessionStatus/AuthSessionStatus';
+import Button from '@/components/Button/Button';
+import GuestLayout from '@/layouts/GuestLayout';
+import Input from '@/components/Input/Input';
+import InputError from '@/components/InputError/InputError';
+import Label from '@/components/Label/Label';
 import Link from 'next/link';
 import { useAuth } from '@/hooks/auth';
 import { useEffect, useState } from 'react';
@@ -53,8 +53,7 @@ const Login = () => {
               <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
             </a>
           </Link>
-        }
-      >
+        }>
         {/* Session Status */}
         <AuthSessionStatus className="mb-4" status={status} />
