+++ src/pages/register.tsx
@@ -1,10 +1,10 @@
-import ApplicationLogo from '@/components/ApplicationLogo';
-import AuthCard from '@/components/AuthCard';
-import Button from '@/components/Button';
-import GuestLayout from '@/components/Layouts/GuestLayout';
-import Input from '@/components/Input';
-import InputError from '@/components/InputError';
-import Label from '@/components/Label';
+import ApplicationLogo from '@/components/ApplicationLogo/ApplicationLogo';
+import AuthCard from '@/components/AuthCard/AuthCard';
+import Button from '@/components/Button/Button';
+import GuestLayout from '@/layouts/GuestLayout';
+import Input from '@/components/Input/Input';
+import InputError from '@/components/InputError/InputError';
+import Label from '@/components/Label/Label';
 import Link from 'next/link';
 import { useAuth } from '@/hooks/auth';
 import { useState } from 'react';
@@ -41,8 +41,7 @@ const Register = () => {
               <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
             </a>
           </Link>
-        }
-      >
+        }>
         <form onSubmit={submitForm}>
           {/* Name */}
           <div>
