+++ src/pages/verify-email.tsx
@@ -1,7 +1,7 @@
-import ApplicationLogo from '@/components/ApplicationLogo';
-import AuthCard from '@/components/AuthCard';
-import Button from '@/components/Button';
-import GuestLayout from '@/components/Layouts/GuestLayout';
+import ApplicationLogo from '@/components/ApplicationLogo/ApplicationLogo';
+import AuthCard from '@/components/AuthCard/AuthCard';
+import Button from '@/components/Button/Button';
+import GuestLayout from '@/layouts/GuestLayout';
 import Link from 'next/link';
 import { useAuth } from '@/hooks/auth';
 import { useState } from 'react';
@@ -23,8 +23,7 @@ const VerifyEmail = () => {
               <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
             </a>
           </Link>
-        }
-      >
+        }>
         <div className="mb-4 text-sm text-gray-600">
           Thanks for signing up! Before getting started, could you verify your email address by
           clicking on the link we just emailed to you? If you didn't receive the email, we will
@@ -46,8 +45,7 @@ const VerifyEmail = () => {
           <button
             type="button"
             className="underline text-sm text-gray-600 hover:text-gray-900"
-            onClick={logout}
-          >
+            onClick={logout}>
             Logout
           </button>
         </div>
